backend/server/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
joblib
numpy
pandas
psycopg2
PyJWT
python-dateutil
pytz
scikit-learn
scipy
six
sqlparse
threadpoolctl
tzdata==
